requirements.txt
fastapi==0.111.1
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
uvicorn[standard]==0.30.3
pytest==8.3.5
